feat(chapter12): add User dataclass and load_from_input function for creating Person instances
